ALGORITMO OPTIMIZADO AL 100 %
Refactored the genetic algorithm and truck assignment logic in Individuo.java to improve modularity and clarity. Added new methods for route planning, truck and warehouse selection, and fitness calculation. Commented out the creation of certain truck types in DataLoader.java and set the genetic algorithm population size to 1 in AlgoritmoGenetico.java. Introduced new utility methods in Gen.java for distance calculation and retrieving the last node in a route.